+++ app/sprinkles/core/src/ServicesProvider/ServicesProvider.php
@@ -240,9 +240,10 @@ class ServicesProvider
                 $config['site.uri.public'] = trim($public, '/');
             }
 
-            // Add asset URLs to the CSRF blacklist
+            // Hacky fix to prevent sessions from being hit too much: ignore CSRF middleware for requests for raw assets ;-)
+            // See https://github.com/laravel/framework/issues/8172#issuecomment-99112012 for more information on why it's bad to hit Laravel sessions multiple times in rapid succession.
             $csrfBlacklist = $config['csrf.blacklist'];
-            $csrfBlacklist['^' . $config['assets.raw.path']] = [
+            $csrfBlacklist['^/' . $config['assets.raw.path']] = [
                 'GET'
             ];