+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## v4.1.8-alpha
+- Normalize paths to always have a leading slash when comparing against the CSRF blacklist (#775)
 - Reimplement `Builder::exclude` to maintain a list of excluded columns, and then automatically update list of columns to fetch in `get()`
 - Deprecate `Model::queryBuilder` and `Model::export`
 - Update nginx config file from spdy to http2